% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{diplomix}
\alias{diplomix}
\title{basic function to do mcmc on the model at a single locus}
\usage{
diplomix(G, reps)
}
\arguments{
\item{G}{an N x 4 matrix (N is the number of individuals) giving
the inferred genotypes of each individual.  The alleles/haplotypes
are coded as integers from 0 up.  For example, an individual might
have the genotype 0031---two doses of the 0 allele, and one each of the
1 and 3 alleles.  Note that the total number of alleles at the locus is
taken to be one greater than the max entry in the genos matrix.  There
can be no missing data in this matrix.}
}
\description{
more later
}
