% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fst_from_ptrace.R
\name{fst_from_ptrace}
\alias{fst_from_ptrace}
\title{compute Fst from the allele frequencies at each rep}
\usage{
fst_from_ptrace(p)
}
\arguments{
\item{p}{a list of matrices, each with 2 rows and A columns where A
is the number of alleles.}
}
\value{
A vector of Fst values.
}
\description{
To be applied to the output of diplomix.  This uses Nei 1973's expression
as grabbed from https://biology.stackexchange.com/questions/40756/calculating-pairwise-fst-from-allele-frequencies
}
