% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{split_pile_alleles}
\alias{split_pile_alleles}
\title{Count the number of alleles in G that are currently allocated to 0 or 1 according to Z}
\usage{
split_pile_alleles(G, Z, out)
}
\arguments{
\item{G}{the N x 4 matrix of allele present in the N individuals.}

\item{Z}{a matrix of 0s and 1s parallel to G.}

\item{out}{a 2 x A NumericMatrix that counts the number of each allele allocated to 0 (first row)
and 1 (second row).  This is numeric so we can add a prior to it. A is the number of alleles.
You have to allocate memory to this outside the function.  The number of columns is important.}
}
\value{
out is an output variable
}
\description{
blah blah
}
\examples{
example(tcf2param_list)
ale_glL <- geno_logL(ale_par_list)
}
\keyword{internal}
