% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tetradiplomix.R
\docType{package}
\name{tetradiplomix}
\alias{tetradiplomix}
\alias{tetradiplomix-package}
\title{tetradiplomix: inference of diploidizing regions of tetraploid genomes}
\description{
tetradiplomix: inference of diploidizing regions of tetraploid genomes
}
