% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_zim}
\alias{update_zim}
\title{update the ZIM matrix}
\usage{
update_zim(G, ZIM, ZM, p)
}
\description{
This updates the allocations for each individuals of the gene copies
to locus 0 or locus 1.  It computes the full conditional for each of the
configurations in ZM, and then it samples from one of those and sets those
values in the row of ZIM.  This only works when called from with RCpp.
Otherwise ZIM does not get updated.
}
