% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{zed_matrix}
\alias{zed_matrix}
\title{Return a 6 x 4 matrix whose rows values of the zeds}
\usage{
zed_matrix()
}
\value{
an integer matrix of 0s and 1s
}
\description{
blah blah
}
\examples{
example(tcf2param_list)
ale_glL <- geno_logL(ale_par_list)
}
\keyword{internal}
